Methods and Functions
Create a program that defines two methods, findLargest and findSmallest. The findLargest method should take an array of integers as input and return the largest value in the array. The findSmallest method should take an array of integers as input and return the smallest value in the array.

Test these methods by calling them from the main method and passing in an array of integers, then printing the result.